# Benchmark and optimizations for style attach