## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Benchmark for attaching styles to controls, this one is testing `TextBox` but we could later easily add more controls. Improvements are small but add up to like 2-3ms for `Calendar` benchmark (~10%) on my machine and reduce memory usage by like 15-20%.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
|         Method |     Mean |    Error |   StdDev |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|--------------- |---------:|---------:|---------:|---------:|---------:|--------:|----------:|
| CreateCalendar | 24.14 ms | 0.337 ms | 0.315 ms | 843.7500 | 406.2500 | 62.5000 |   4.84 MB |

|              Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
| AttachTextBoxStyles | 23.90 us | 0.227 us | 0.189 us | 2.1667 | 0.4272 |     - |  10.34 KB |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
|         Method |     Mean |    Error |   StdDev |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|--------------- |---------:|---------:|---------:|---------:|---------:|--------:|----------:|
| CreateCalendar | 21.10 ms | 0.383 ms | 0.377 ms | 718.7500 | 343.7500 | 31.2500 |   4.17 MB |

|              Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
| AttachTextBoxStyles | 19.02 us | 0.164 us | 0.153 us | 1.9531 | 0.6104 |     - |   8.06 KB |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
General case of LINQ and enumerator boxing removal.

## Checklist

- [x] Added ~~unit tests~~ benchmarks (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation