@MarchingCube 2020-02-12
I need to run benchmarks again to get results and then I will finalize this PR.